Move support code for GothamService to `service`
